% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pwgarch.R
\name{treatments}
\alias{treatments}
\title{View individual series treatments of the prewhitening scheme}
\usage{
treatments(x)
}
\arguments{
\item{x}{a data.frame as returned from \code{pwgarch}}
}
\value{
nothing, invoked for side effects (printing)
}
\description{
View individual series treatments of the prewhitening scheme
}
